% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishr-package.R
\docType{package}
\name{fishr-package}
\alias{fishr}
\alias{fishr-package}
\title{fishr: Calculate standard catch metrics for fisheries}
\description{
Provides functions for calculating and analyzing fisheries catch data, such as Catch Per Unit Effort (CPUE), a fundamental metric in fisheries science.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ateucher/fishr.example}
  \item Report bugs at \url{https://github.com/ateucher/fishr.example/issues}
}

}
\author{
\strong{Maintainer}: Andy Teucher \email{andy.teucher@gmail.com} (\href{https://orcid.org/0000-0002-7840-692X}{ORCID})

}
\keyword{internal}
